class 24 finished

+++ app/Http/Controllers/MoyusController.php
@@ -111,6 +111,11 @@ class MoyusController extends Controller
      */
     public function destroy($channel, Moyu $moyu)
     {
+        $this->authorize('update', $moyu);
+        if($moyu->user_id != auth()->id()){
+          abort(403, 'You do not have the permission to do this!');
+        }
+
         $moyu->delete();
 
         if(request()->wantsJson()){
